@@ -7,85 +7,86 @@
   "prefer-stable": true,
   "autoload": {
     "psr-4": {
-      "App\\": "src"
+      "Surfnet\\Tiqr\\": "src",
+      "Surfnet\\Tiqr\\Dev\\": "dev"
     }
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "Dev\\": "dev",
-      "Tests\\": "tests/"
-    },
-    "files": [
-      "vendor/symfony/var-dumper/Resources/functions/dump.php"
-    ]
   },
   "require": {
     "ext-dom": "*",
     "ext-libxml": "*",
     "ext-zlib": "*",
-    "openconext/monitor-bundle": "^2.1",
-    "sensio/framework-extra-bundle": "^5.4",
-    "surfnet/stepup-bundle": "^4.1",
-    "surfnet/stepup-gssp-bundle": "^4",
-    "surfnet/stepup-saml-bundle": "^4.3",
-    "symfony/asset": "^4.3",
-    "symfony/console": "^4.3",
-    "symfony/dotenv": "^4.3",
-    "symfony/flex": "^1.4",
-    "symfony/form": "^4.3",
-    "symfony/monolog-bundle": "^3.4",
-    "symfony/security-bundle": "^4.3",
-    "symfony/translation": "^4.3",
-    "symfony/twig-bundle": "^4.3",
-    "symfony/validator": "^4.3",
-    "symfony/webpack-encore-bundle": "^1.6",
-    "symfony/yaml": "^4.3",
-    "tiqr/tiqr-server-libphp": "^3.0"
+    "incenteev/composer-parameter-handler": "*",
+    "openconext/monitor-bundle": "^4.0",
+    "surfnet/stepup-bundle": "^6.0.9",
+    "surfnet/stepup-gssp-bundle": "^5.0",
+    "surfnet/stepup-saml-bundle": "^6.0",
+    "symfony/asset": "6.4.*",
+    "symfony/config": "^6.4",
+    "symfony/console": "6.4.*",
+    "symfony/dependency-injection": "^6.4",
+    "symfony/flex": "^2.4",
+    "symfony/form": "6.4.*",
+    "symfony/monolog-bundle": "^3.8",
+    "symfony/runtime": "6.*.*",
+    "symfony/security-bundle": "6.4.*",
+    "symfony/templating": "^6.4",
+    "symfony/translation": "^6.4",
+    "symfony/twig-bundle": "6.4.*",
+    "symfony/validator": "6.4.*",
+    "symfony/webpack-encore-bundle": "^2.1",
+    "symfony/yaml": "6.4.*",
+    "tiqr/tiqr-server-libphp": "^4.0",
+    "twig/extra-bundle": "^3.0",
+    "twig/twig": "^3.0"
   },
   "require-dev": {
-    "behat/mink": "^1.7",
-    "behat/mink-browserkit-driver": "^1.3",
-    "behat/mink-goutte-driver": "^1.2",
-    "behat/symfony2-extension": "^2.1",
-    "behatch/contexts": "^3.2",
-    "jakub-onderka/php-parallel-lint": "~1",
-    "khanamiryan/qrcode-detector-decoder": "^1.0.3",
-    "league/csv": "^8.2",
-    "malukenho/docheader": "^0",
-    "mockery/mockery": "^1",
-    "phpmd/phpmd": "~2",
-    "phpunit/phpcov": "~6",
-    "phpunit/phpunit": "^8.5",
-    "squizlabs/php_codesniffer": "~3",
-    "symfony/browser-kit": "^4.3",
-    "symfony/css-selector": "^3",
-    "symfony/maker-bundle": "~1",
-    "symfony/phpunit-bridge": "^5.3",
-    "symfony/stopwatch": "^5.3",
-    "symfony/web-profiler-bundle": "^5.0"
+    "behat/mink": "^1.9",
+    "behat/mink-browserkit-driver": "^2.1",
+    "behat/mink-selenium2-driver": "^1.6",
+    "friends-of-behat/symfony-extension": "^2.4",
+    "guzzle/guzzle": "^3.8",
+    "khanamiryan/qrcode-detector-decoder": "^2.0",
+    "league/csv": "^9.13",
+    "malukenho/docheader": "^1",
+    "mockery/mockery": "^1.6",
+    "overtrue/phplint": "*",
+    "phpmd/phpmd": "^2.15",
+    "phpstan/phpstan": "^1.10",
+    "phpstan/phpstan-symfony": "^1.3",
+    "phpunit/phpunit": "^9.6",
+    "sebastian/phpcpd": "^6.0",
+    "slevomat/coding-standard": "^8.13",
+    "soyuka/contexts": "^3.3",
+    "squizlabs/php_codesniffer": "^3.7",
+    "symfony/browser-kit": "6.4.*",
+    "symfony/css-selector": "^6.4",
+    "symfony/phpunit-bridge": "6.4.*",
+    "symfony/stopwatch": "6.4.*",
+    "symfony/web-profiler-bundle": "6.4.*"
   },
   "scripts": {
-    "test": [
+    "check": [
+      "@validate",
+      "@docheader",
       "@lint",
-      "@static-analysis",
-      "@phpunit",
-      "@behat",
-      "@jest",
-      "@security-tests"
+      "@jscpd",
+      "@phpcs",
+      "@phpmd",
+      "@phpstan",
+      "@unit-tests",
+      "@jest"
     ],
-      "lint": "./ci/qa/lint",
-      "static-analysis": "./ci/qa/static-analysis",
-      "phpunit": "./ci/qa/phpunit",
-      "behat": "./ci/qa/behat",
-      "jest": "./ci/qa/jest",
-      "security-tests": "./ci/qa/security-tests",
-
-    "coverage": [
-      "@phpunit-coverage",
-      "mkdir -p coverage/reports",
-      "vendor/bin/phpcov merge coverage/reports --html coverage"
-    ],
-    "phpunit-coverage": "vendor/bin/phpunit --coverage-php coverage/reports/unit.cov",
+    "composer-validate": "./ci/qa/validate",
+    "docheader": "./ci/qa/docheader",
+    "lint": "./ci/qa/lint",
+    "phpcs": "./ci/qa/phpcs",
+    "phpcbf": "./ci/qa/phpcbf",
+    "phpmd": "./ci/qa/phpmd",
+    "phpstan": "./ci/qa/phpstan",
+    "phpstan-baseline": "./ci/qa/phpstan-update-baseline",
+    "unit-tests": "./ci/qa/phpunit",
+    "behat": "./ci/qa/behat",
+    "jest": "./ci/qa/jest",
     "encore": [
       "yarn encore production"
     ],
@@ -93,25 +94,43 @@
       "yarn install --frozen-lockfile",
       "yarn encore production"
     ],
-    "post-install-cmd": [
-        "@auto-scripts"
-    ],
-    "post-update-cmd": [
-        "@auto-scripts"
-    ],
     "auto-scripts": {
       "cache:clear": "symfony-cmd",
       "assets:install %PUBLIC_DIR%": "symfony-cmd"
-    }
+    },
+    "post-install-cmd": [
+      "Incenteev\\ParameterHandler\\ScriptHandler::buildParameters",
+      "@auto-scripts"
+    ],
+    "post-update-cmd": [
+      "Incenteev\\ParameterHandler\\ScriptHandler::buildParameters",
+      "@auto-scripts"
+    ]
   },
   "config": {
     "sort-packages": true,
-    "allow-plugins": {
-      "symfony/flex": true
-    },
     "platform": {
-      "php": "7.2.5"
+      "php": "8.2"
+    },
+    "allow-plugins": {
+      "symfony/flex": true,
+      "symfony/runtime": true,
+      "dealerdirect/phpcodesniffer-composer-installer": true
     }
+  },
+  "extra": {
+    "runtime": {
+      "disable_dotenv": true
+    },
+    "symfony": {
+      "allow-contrib": false,
+      "require": "6.4.*"
+    },
+    "incenteev-parameters": [
+      {
+        "file": "config/openconext/parameters.yaml"
+      }
+    ]
   },
   "archive": {
     "exclude": [
