@@ -15,34 +15,32 @@
  * limitations under the License.
  */
 
-namespace Dev\Controller;
+declare(strict_types = 1);
 
-use App\Tiqr\TiqrServiceInterface;
-use Symfony\Component\Routing\Annotation\Route;
+namespace Surfnet\Tiqr\Dev\Controller;
+
+use InvalidArgumentException;
+use Surfnet\Tiqr\Tiqr\TiqrServiceInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\StreamedResponse;
+use Symfony\Component\Routing\Attribute\Route;
 
 /**
  * Exposes the qr codes as links.
  */
 final class QrLinkController extends AbstractController
 {
-    private $tiqrService;
-
-    public function __construct(TiqrServiceInterface $tiqrService)
+    public function __construct(private readonly TiqrServiceInterface $tiqrService)
     {
-        $this->tiqrService = $tiqrService;
     }
 
     /**
      * Returns the QR for registration without an active authNRequest.
-     *
-     * @Route("/registration/qr/dev", name="app_identity_registration_qr_dev", methods={"GET"})
-     * @param Request $request
-     * @return \Symfony\Component\HttpFoundation\StreamedResponse
      */
-    public function registrationQrAction(Request $request)
+    #[Route(path: '/registration/qr/dev', name: 'app_identity_registration_qr_dev', methods: ['GET'])]
+    public function registrationQr(Request $request): StreamedResponse
     {
         $key = $this->tiqrService->generateEnrollmentKey('dev');
         $metadataURL = $request->getUriForPath(sprintf('/tiqr.php?key=%s', urlencode($key)));
@@ -51,13 +49,10 @@ final class QrLinkController extends AbstractController
 
     /**
      * Returns the link for registration without an active authNRequest.
-     *
-     * @Route("/registration/qr/link", name="app_identity_registration_qr_link", methods={"GET"})
-     * @param Request $request
-     * @return Response
-     * @throws \InvalidArgumentException
+     * @throws InvalidArgumentException
      */
-    public function registrationLinkAction(Request $request)
+    #[Route(path: '/registration/qr/link', name: 'app_identity_registration_qr_link', methods: ['GET'])]
+    public function registrationLink(Request $request): Response
     {
         $key = $this->tiqrService->generateEnrollmentKey('dev');
         $metadataUrl = $request->getUriForPath(sprintf('/tiqr.php?key=%s', urlencode($key)));
@@ -73,12 +68,9 @@ final class QrLinkController extends AbstractController
 
     /**
      * Returns the QR without an active authNRequest.
-     *
-     * @Route("/authentication/qr/{nameId}", name="app_identity_authentication_qr_dev", methods={"GET"})
-     * @param string $nameId
-     * @return \Symfony\Component\HttpFoundation\StreamedResponse
      */
-    public function authenticationQrAction($nameId)
+    #[Route(path: '/authentication/qr/{nameId}', name: 'app_identity_authentication_qr_dev', methods: ['GET'])]
+    public function authenticationQr(string $nameId): StreamedResponse
     {
         $this->tiqrService->startAuthentication($nameId, 'dev');
         return $this->tiqrService->createAuthenticationQRResponse();
@@ -86,12 +78,9 @@ final class QrLinkController extends AbstractController
 
     /**
      * Returns the link without an active authNRequest.
-     *
-     * @Route("/authentication/qr/{nameId}/link", name="app_identity_authentication_qr_link", methods={"GET"})
-     * @param string $nameId
-     * @return Response
      */
-    public function authenticationQrLinkAction($nameId)
+    #[Route(path: '/authentication/qr/{nameId}/link', name: 'app_identity_authentication_qr_link', methods: ['GET'])]
+    public function authenticationQrLink(string $nameId): Response
     {
         $this->tiqrService->startAuthentication($nameId, 'dev');
         $challengeUrl = $this->tiqrService->authenticationUrl();
