@@ -16,58 +16,54 @@
  * limitations under the License.
  */
 
-declare(strict_types=1);
+declare(strict_types = 1);
 
-namespace Dev\Controller;
+namespace Surfnet\Tiqr\Dev\Controller;
 
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Routing\Annotation\Route;
 use DOMDocument;
 use RobRichards\XMLSecLibs\XMLSecurityKey;
-use SAML2\Assertion;
 use SAML2\Certificate\PrivateKeyLoader;
 use SAML2\Configuration\PrivateKey;
 use SAML2\DOMDocumentFactory;
 use SAML2\Message;
 use SAML2\Response;
-use Surfnet\GsspBundle\Service\RegistrationService;
 use Surfnet\SamlBundle\Entity\IdentityProvider;
 use Surfnet\SamlBundle\Entity\ServiceProvider;
 use Surfnet\SamlBundle\Http\Exception\AuthnFailedSamlResponseException;
 use Surfnet\SamlBundle\Http\PostBinding;
 use Surfnet\SamlBundle\SAML2\AuthnRequest;
 use Surfnet\SamlBundle\SAML2\AuthnRequestFactory;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response as SymfonyResponse;
+use Symfony\Component\Routing\Attribute\Route;
 
 /**
  * Demo SP.
+ * These controller actions aid developers and testers with adding a test/demo
+ * authentication and registration feature. This enables the tester/developer
+ * to test the GSSP component without installing an SP that interacts with the
+ * GSSP.
+ *
+ * Note that this controller should never be enabled in production.
  */
 final class SPController extends AbstractController
 {
-    private $identityProvider;
-    private $serviceProvider;
-    private $postBinding;
-
     public function __construct(
-        ServiceProvider $serviceProvider,
-        IdentityProvider $identityProvider,
-        PostBinding $postBinding
+        private readonly ServiceProvider $serviceProvider,
+        private readonly IdentityProvider $identityProvider,
+        private readonly PostBinding $postBinding
     ) {
-        $this->identityProvider = $identityProvider;
-        $this->serviceProvider = $serviceProvider;
-        $this->postBinding = $postBinding;
     }
 
     /**
-     * @Route("/demo/sp", name="sp_demo")
-     * @Route("/", name="homepage")
-     *
-     * See @see RegistrationService for a more clean example.
      *
      * @throws \Exception
      */
-    public function demoSpAction(Request $request)
+    #[Route(path: '/demo/sp', name: 'sp_demo')]
+    #[Route(path: '/', name: 'homepage')]
+    public function demoSp(Request $request): SymfonyResponse
     {
         if (!$request->isMethod(Request::METHOD_POST)) {
             return $this->render('dev/sp.html.twig', ['nameId' => $request->get('nameId')]);
@@ -102,28 +98,23 @@ final class SPController extends AbstractController
         return $response;
     }
 
-    /**
-     * @Route("/demo/sp/acs", name="sp_demo_acs")
-     *
-     * See @see RegistrationService for a more clean example.
-     */
-    public function assertionConsumerServiceAction(Request $request)
+    #[Route(path: '/demo/sp/acs', name: 'sp_demo_acs')]
+    public function assertionConsumerService(Request $request): SymfonyResponse
     {
         $xmlResponse = $request->request->get('SAMLResponse');
         $xml = base64_decode($xmlResponse);
         try {
-            /** @var Assertion $response */
             $response = $this->postBinding->processResponse($request, $this->identityProvider, $this->serviceProvider);
 
             $nameID = $response->getNameId();
 
             return $this->render('dev/acs.html.twig', [
                 'requestId' => $response->getId(),
-                'nameId' => $nameID ? [
-                    'value' => $nameID->value,
-                    'format' => $nameID->Format,
+                'nameId' => $nameID !== null ? [
+                    'value' => $nameID->getValue(),
+                    'format' => $nameID->getFormat(),
                 ] : [],
-                'issuer' => $response->getIssuer(),
+                'issuer' => $response->getIssuer()->getValue(),
                 'relayState' => $request->get(AuthnRequest::PARAMETER_RELAY_STATE, ''),
                 'authenticatingAuthority' => $response->getAuthenticatingAuthority(),
                 'xml' => $this->toFormattedXml($xml),
@@ -135,7 +126,7 @@ final class SPController extends AbstractController
                 'error' => $e->getMessage(),
                 'status' => $samlResponse->getStatus(),
                 'requestId' => $samlResponse->getId(),
-                'issuer' => $samlResponse->getIssuer(),
+                'issuer' => $samlResponse->getIssuer()->getValue(),
                 'relayState' => $request->get(AuthnRequest::PARAMETER_RELAY_STATE, ''),
                 'xml' => $this->toFormattedXml($xml),
             ]);
@@ -144,12 +135,8 @@ final class SPController extends AbstractController
 
     /**
      * Formats xml.
-     *
-     * @param string $xml
-     *
-     * @return string
      */
-    private function toFormattedXml($xml)
+    private function toFormattedXml(string|bool $xml): string|false
     {
         $domxml = new DOMDocument('1.0');
         $domxml->preserveWhiteSpace = false;
@@ -161,13 +148,9 @@ final class SPController extends AbstractController
 
     /**
      * Sign AuthnRequest query parameters.
-     *
-     * @param array $queryParams
-     * @return string
-     *
      * @throws \Exception
      */
-    private function signRequestQuery(array $queryParams)
+    private function signRequestQuery(array $queryParams): string
     {
         /** @var  $securityKey */
         $securityKey = $this->loadServiceProviderPrivateKey();
@@ -175,7 +158,7 @@ final class SPController extends AbstractController
         $toSign = http_build_query($queryParams);
         $signature = $securityKey->signData($toSign);
 
-        return $toSign.'&Signature='.urlencode(base64_encode($signature));
+        return $toSign.'&Signature='.urlencode(base64_encode((string) $signature));
     }
 
     /**
@@ -185,7 +168,7 @@ final class SPController extends AbstractController
      *
      * @throws \Exception
      */
-    private function loadServiceProviderPrivateKey()
+    private function loadServiceProviderPrivateKey(): XMLSecurityKey
     {
         $keyLoader = new PrivateKeyLoader();
         $privateKey = $keyLoader->loadPrivateKey(
@@ -198,18 +181,13 @@ final class SPController extends AbstractController
     }
 
     /**
-     * @param string $xml
      *
      * @return Message
-     *
      * @throws \Exception
      */
-    private function toUnsignedErrorResponse($xml)
+    private function toUnsignedErrorResponse(string $xml): Message
     {
-        $previous = libxml_disable_entity_loader(true);
         $asXml = DOMDocumentFactory::fromString($xml);
-        libxml_disable_entity_loader($previous);
-
         return Response::fromXML($asXml->documentElement);
     }
 }
