@@ -15,7 +15,9 @@
  * limitations under the License.
  */
 
-namespace Dev\Twig;
+declare(strict_types = 1);
+
+namespace Surfnet\Tiqr\Dev\Twig;
 
 use Surfnet\SamlBundle\Entity\HostedEntities;
 use Twig\Extension\AbstractExtension;
@@ -23,21 +25,16 @@ use Twig\TwigFunction;
 
 final class GsspExtension extends AbstractExtension
 {
-    private $hostedEntities;
-
-    public function __construct(HostedEntities $hostedEntities)
+    public function __construct(private readonly HostedEntities $hostedEntities)
     {
-        $this->hostedEntities = $hostedEntities;
     }
 
     public function getFunctions()
     {
-        return array(
-            new TwigFunction('demoSpUrl', array($this, 'generateDemoSPUrl')),
-        );
+        return [new TwigFunction('demoSpUrl', $this->generateDemoSPUrl(...))];
     }
 
-    public function generateDemoSPUrl()
+    public function generateDemoSPUrl(): string
     {
         return sprintf(
             'https://pieter.aai.surfnet.nl/simplesamlphp/sp.php?idp=%s',
