@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types = 1);
+
 /**
  * Copyright 2018 SURFnet B.V.
  *
@@ -16,33 +18,23 @@
  * limitations under the License.
  */
 
-namespace App\Service;
+namespace Surfnet\Tiqr\Service;
 
 use Symfony\Component\HttpFoundation\Request;
 
 final class UserAgentMatcher implements UserAgentMatcherInterface
 {
-    /**
-     * @var string
-     */
-    private $pattern;
-
-    /**
-     * @param string $pattern
-     */
-    public function __construct($pattern)
+    public function __construct(private readonly string $pattern)
     {
-        $this->pattern = $pattern;
     }
 
-    /**
-     * @param Request $request
-     * @return bool
-     */
-    public function isOfficialTiqrMobileApp(Request $request)
+    public function isOfficialTiqrMobileApp(Request $request): bool
     {
         $userAgent = $request->headers->get('User-Agent');
-
-        return preg_match($this->pattern, $userAgent);
+        if (!$userAgent) {
+            return false;
+        }
+        $result = preg_match($this->pattern, $userAgent);
+        return $result === 1;
     }
 }
