@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types = 1);
+
 /**
  * Copyright 2018 SURFnet B.V.
  *
@@ -15,17 +18,18 @@
  * limitations under the License.
  */
 
-namespace App\Tiqr\Legacy;
+namespace Surfnet\Tiqr\Tiqr\Legacy;
 
-use App\Exception\TiqrServerRuntimeException;
-use App\Tiqr\Response\AuthenticationErrorResponse;
-use App\Tiqr\Response\AuthenticationResponse;
-use App\Tiqr\Response\RejectedAuthenticationResponse;
-use App\Tiqr\Response\ValidAuthenticationResponse;
-use App\Tiqr\TiqrServiceInterface;
-use App\Tiqr\TiqrUserInterface;
 use Exception;
 use Psr\Log\LoggerInterface;
+use Surfnet\Tiqr\Exception\TiqrServerRuntimeException;
+use Surfnet\Tiqr\Tiqr\Response\AuthenticationErrorResponse;
+use Surfnet\Tiqr\Tiqr\Response\AuthenticationResponse;
+use Surfnet\Tiqr\Tiqr\Response\RejectedAuthenticationResponse;
+use Surfnet\Tiqr\Tiqr\Response\ValidAuthenticationResponse;
+use Surfnet\Tiqr\Tiqr\TiqrServiceInterface;
+use Surfnet\Tiqr\Tiqr\TiqrUserInterface;
+use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Component\HttpFoundation\StreamedResponse;
 use Tiqr_Service;
@@ -35,36 +39,22 @@ use Tiqr_StateStorage_StateStorageInterface;
  * Wrapper around the legacy Tiqr service.
  *
  * @SuppressWarnings(PHPMD.TooManyPublicMethods)
+ * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
  * It's Legacy.
  */
 final class TiqrService implements TiqrServiceInterface
 {
-    const ENROLL_KEYS_SESSION_NAME = 'enrollment-session-keys';
-    /**
-     * @var \Tiqr_Service
-     */
-    private $tiqrService;
-    private $tiqrStateStorage;
-    private $session;
-    private $logger;
-
-    /**
-     * @var string
-     */
-    private $accountName;
+    public const ENROLL_KEYS_SESSION_NAME = 'enrollment-session-keys';
+    private SessionInterface $session;
 
     public function __construct(
-        Tiqr_Service $tiqrService,
-        Tiqr_StateStorage_StateStorageInterface $tiqrStateStorage,
-        SessionInterface $session,
-        LoggerInterface $logger,
-        $accountName
+        private readonly Tiqr_Service $tiqrService,
+        private readonly Tiqr_StateStorage_StateStorageInterface $tiqrStateStorage,
+        private readonly RequestStack $requestStack,
+        private readonly LoggerInterface $logger,
+        private readonly string $appSecret,
+        private readonly string $accountName
     ) {
-        $this->tiqrService = $tiqrService;
-        $this->tiqrStateStorage = $tiqrStateStorage;
-        $this->session = $session;
-        $this->logger = $logger;
-        $this->accountName = $accountName;
     }
 
     /**
@@ -73,7 +63,7 @@ final class TiqrService implements TiqrServiceInterface
     public function createRegistrationQRResponse(string $metadataURL): StreamedResponse
     {
         try {
-            return new StreamedResponse(function () use ($metadataURL) {
+            return new StreamedResponse(function () use ($metadataURL): void {
                 $this->tiqrService->generateEnrollmentQR($metadataURL);
             });
         } catch (Exception $e) {
@@ -101,6 +91,8 @@ final class TiqrService implements TiqrServiceInterface
      */
     public function generateEnrollmentKey(string $sari): string
     {
+        $this->initSession();
+
         $this->logger->debug('Generating userId');
         $userId = $this->generateId();
         $this->logger->debug('Storing the userId to session state');
@@ -130,6 +122,7 @@ final class TiqrService implements TiqrServiceInterface
      */
     private function storeEnrollmentKey(string $key): void
     {
+        $this->initSession();
         $keys = [];
         if ($this->session->has(self::ENROLL_KEYS_SESSION_NAME)) {
             $keys = $this->session->get(self::ENROLL_KEYS_SESSION_NAME);
@@ -140,6 +133,7 @@ final class TiqrService implements TiqrServiceInterface
 
     /**
      * @see TiqrServiceInterface::getEnrollmentMetadata()
+     * @return array<string, mixed>
      */
     public function getEnrollmentMetadata(string $key, string $loginUri, string $enrollmentUrl): array
     {
@@ -182,6 +176,7 @@ final class TiqrService implements TiqrServiceInterface
      */
     public function startAuthentication(string $userId, string $sari): string
     {
+        $this->initSession();
         try {
             $sessionKey = $this->tiqrService->startAuthenticationSession($userId, $this->session->getId());
             $this->session->set('sessionKey', $sessionKey);
@@ -200,6 +195,7 @@ final class TiqrService implements TiqrServiceInterface
      */
     public function isAuthenticated(): bool
     {
+        $this->initSession();
         return $this->tiqrService->getAuthenticatedUser($this->session->getId()) !== null;
     }
 
@@ -214,7 +210,7 @@ final class TiqrService implements TiqrServiceInterface
     {
         $sessionKey = $this->getAuthenticationSessionKey();
         try {
-            return new StreamedResponse(function () use ($sessionKey) {
+            return new StreamedResponse(function () use ($sessionKey): void {
                 $this->tiqrService->generateAuthQR($sessionKey);
             });
         } catch (Exception $e) {
@@ -247,20 +243,14 @@ final class TiqrService implements TiqrServiceInterface
     {
         try {
             $result = $this->tiqrService->authenticate($user->getId(), $user->getSecret(), $sessionKey, $response);
-            switch ($result) {
-                case Tiqr_Service::AUTH_RESULT_AUTHENTICATED:
-                    return new ValidAuthenticationResponse('OK');
-                case Tiqr_Service::AUTH_RESULT_INVALID_CHALLENGE:
-                    return new AuthenticationErrorResponse('INVALID_CHALLENGE');
-                case Tiqr_Service::AUTH_RESULT_INVALID_REQUEST:
-                    return new AuthenticationErrorResponse('INVALID_REQUEST');
-                case Tiqr_Service::AUTH_RESULT_INVALID_RESPONSE:
-                    return new RejectedAuthenticationResponse('INVALID_RESPONSE');
-                case Tiqr_Service::AUTH_RESULT_INVALID_USERID:
-                    return new AuthenticationErrorResponse('INVALID_USER');
-                default:
-                    return new AuthenticationErrorResponse('ERROR');
-            }
+            return match ($result) {
+                Tiqr_Service::AUTH_RESULT_AUTHENTICATED => new ValidAuthenticationResponse('OK'),
+                Tiqr_Service::AUTH_RESULT_INVALID_CHALLENGE => new AuthenticationErrorResponse('INVALID_CHALLENGE'),
+                Tiqr_Service::AUTH_RESULT_INVALID_REQUEST => new AuthenticationErrorResponse('INVALID_REQUEST'),
+                Tiqr_Service::AUTH_RESULT_INVALID_RESPONSE => new RejectedAuthenticationResponse('INVALID_RESPONSE'),
+                Tiqr_Service::AUTH_RESULT_INVALID_USERID => new AuthenticationErrorResponse('INVALID_USER'),
+                default => new AuthenticationErrorResponse('ERROR'),
+            };
         } catch (Exception $e) {
             $this->logger->error(
                 sprintf('Error authenticating user "%s": %s', $user->getId(), $e->getMessage())
@@ -275,6 +265,7 @@ final class TiqrService implements TiqrServiceInterface
      */
     public function getEnrollmentStatus(): int
     {
+        $this->initSession();
         try {
             return $this->tiqrService->getEnrollmentStatus($this->session->getId());
         } catch (Exception $e) {
@@ -299,6 +290,7 @@ final class TiqrService implements TiqrServiceInterface
      */
     public function getUserId(): string
     {
+        $this->initSession();
         return $this->session->get('userId');
     }
 
@@ -307,6 +299,7 @@ final class TiqrService implements TiqrServiceInterface
      */
     public function getAuthenticationSessionKey(): string
     {
+        $this->initSession();
         return $this->session->get('sessionKey');
     }
 
@@ -320,7 +313,7 @@ final class TiqrService implements TiqrServiceInterface
             if (false === $translatedAddress) {
                 throw new TiqrServerRuntimeException(sprintf('Error translating address for "%s"', $notificationAddress));
             }
-            $this->tiqrService->sendAuthNotification($this->getAuthenticationSessionKey(), $notificationType, $translatedAddress);
+            $this->tiqrService->sendAuthNotification($this->getAuthenticationSessionKey(), $notificationType, (string) $translatedAddress);
         } catch (Exception $e) {
             throw TiqrServerRuntimeException::fromOriginalException($e);
         }
@@ -329,37 +322,33 @@ final class TiqrService implements TiqrServiceInterface
 
     //////////
     // Private
-
     /**
      * Currently the legacy way to generate the user Tiqr id.
      *
      * TODO:maybe use something like UUID?
      *
-     * @param int $length
      *
      * @return string
      */
     private function generateId(int $length = 4): string
     {
-        return base_convert(time(), 10, 36) . '-' . base_convert(mt_rand(0, pow(36, $length)), 10, 36);
+        return base_convert((string) time(), 10, 36) . '-' . base_convert((string) mt_rand(0, 36 ** $length), 10, 36);
     }
 
     /** Create a stable hash from $identifier
-     * @param string $identifier
      * @return string The hashed version of $identifier
      */
     private function getHashedIdentifier(string $identifier): string
     {
-        // TODO: Use Symfony configuration system instead of accessing $_ENV directly
-        return hash_hmac('sha256', $identifier, $_ENV['APP_SECRET']);
+        return hash_hmac('sha256', $identifier, $this->appSecret);
     }
 
     public function getSariForSessionIdentifier(string $identifier): string
     {
         try {
-            $hashed_identifier = $this->getHashedIdentifier($identifier);
-            $res = $this->tiqrStateStorage->getValue('sari_' . $hashed_identifier);
-        } catch (Exception $e) {
+            $hashedIdentifier = $this->getHashedIdentifier($identifier);
+            $res = $this->tiqrStateStorage->getValue('sari_' . $hashedIdentifier);
+        } catch (Exception) {
             $this->logger->error(sprintf('Error getting SARI for identifier "%s"', substr($identifier, 0, 8)));
             return '';
         }
@@ -369,30 +358,6 @@ final class TiqrService implements TiqrServiceInterface
         }
 
         return $res;
-    }
-
-    /**
-     * @param string $identifier Enrollment key or session key
-     * @return bool true on success, false otherwise
-     * Does not throw
-     */
-    private function unsetSariForSessionIdentifier(string $identifier): bool
-    {
-        $this->logger->info(
-            sprintf("Unsetting SARI for identifier '%s...'", substr($identifier, 0, 8))
-        );
-
-        try {
-            $hashed_identifier = $this->getHashedIdentifier($identifier);
-            $this->tiqrStateStorage->unsetValue('sari_' . $hashed_identifier);
-        } catch (Exception $e) {
-            $this->logger->error(
-                sprintf('unsetSariForSessionIdentifier failed for "%s"', substr($identifier, 0, 8))
-            );
-            return false;
-        }
-
-        return true;
     }
 
     /**
@@ -422,8 +387,8 @@ final class TiqrService implements TiqrServiceInterface
             sprintf("Setting SARI '%s' for identifier '%s...'", $sari, substr($identifier, 0, 8))
         );
         try {
-            $hashed_identifier = $this->getHashedIdentifier($identifier);
-            $this->tiqrStateStorage->setValue('sari_' . $hashed_identifier, $sari, 60 * 60);
+            $hashedIdentifier = $this->getHashedIdentifier($identifier);
+            $this->tiqrStateStorage->setValue('sari_' . $hashedIdentifier, $sari, 60 * 60);
         } catch (Exception $e) {
             // Catch errors from the tiqr-server and up-cycle them to  exceptions that are meaningful to our domain
             throw TiqrServerRuntimeException::fromOriginalException($e);
@@ -436,6 +401,8 @@ final class TiqrService implements TiqrServiceInterface
      */
     private function clearPreviousEnrollmentState(): void
     {
+        $this->initSession();
+
         $keys = [];
         if ($this->session->has(self::ENROLL_KEYS_SESSION_NAME)) {
             $keys = $this->session->get(self::ENROLL_KEYS_SESSION_NAME);
@@ -455,5 +422,10 @@ final class TiqrService implements TiqrServiceInterface
         $this->logger->debug(sprintf($format, count($keys)));
 
         $this->session->set(self::ENROLL_KEYS_SESSION_NAME, $keys);
+    }
+
+    private function initSession(): void
+    {
+        $this->session = $this->requestStack->getSession();
     }
 }
