@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types = 1);
+
 /**
  * Copyright 2018 SURFnet B.V.
  *
@@ -15,18 +18,15 @@
  * limitations under the License.
  */
 
-namespace App\Tiqr\Response;
+namespace Surfnet\Tiqr\Tiqr\Response;
 
 /**
  * Response when the user has given a wrong password.
  */
 class RejectedAuthenticationResponse implements AuthenticationResponse
 {
-    private $error;
-
-    public function __construct(string $error)
+    public function __construct(private readonly string $error)
     {
-        $this->error = $error;
     }
 
     /**
