@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types = 1);
+
 /**
  * Copyright 2018 SURFnet B.V.
  *
@@ -15,36 +18,39 @@
  * limitations under the License.
  */
 
-namespace App;
+namespace Surfnet\Tiqr;
 
 use Psr\Log\AbstractLogger;
 use Psr\Log\LoggerInterface;
+use Stringable;
 
 final class WithContextLogger extends AbstractLogger
 {
-    private $logger;
-    private $context;
-
-    private function __construct(LoggerInterface $logger, array $context)
-    {
-        $this->logger = $logger;
-        $this->context = $context;
+    private function __construct(
+        private readonly LoggerInterface $logger,
+        /** @var array<string, string> $context */
+        private readonly array $context
+    ) {
     }
 
-    public static function from(LoggerInterface $logger, array $context)
-    {
+    /**
+     * @param array<string,string> $context
+     */
+    public static function from(
+        LoggerInterface $logger,
+        array $context
+    ): self {
         return new self($logger, $context);
     }
 
     /**
      * Logs with an arbitrary level.
      *
-     * @param mixed $level
-     * @param array $context
+     * @param array<string, string> $context
      *
      * @return void
      */
-    public function log($level, $message, array $context = array())
+    public function log($level, string|Stringable $message, array $context = []): void
     {
         $this->logger->log($level, $message, array_merge($this->context, $context));
     }
