@@ -1,23 +1,29 @@
 <?php
 
-use Symfony\Component\Dotenv\Dotenv;
+use Symfony\Component\Yaml\Yaml;
 
 require dirname(__DIR__).'/vendor/autoload.php';
+$parametersPath = dirname(__DIR__).'/config/openconext/parameters.yaml';
+$parameters = Yaml::parseFile($parametersPath);
+$parameters = $parameters['parameters'];
+$requiredParameters = ['app_env', 'app_debug', 'app_secret'];
 
-// Load cached env vars if the .env.local.php file exists
-// Run "composer dump-env prod" to create it (requires symfony/flex >=1.2)
-if (is_array($env = @include dirname(__DIR__).'/.env.local.php')) {
-    foreach ($env as $k => $v) {
-        $_ENV[$k] = $_ENV[$k] ?? (isset($_SERVER[$k]) && 0 !== strpos($k, 'HTTP_') ? $_SERVER[$k] : $v);
-    }
-} elseif (!class_exists(Dotenv::class)) {
-    throw new RuntimeException('Please run "composer require symfony/dotenv" to load the ".env" files configuring the application.');
-} else {
-    // load all the .env files
-    (new Dotenv(false))->loadEnv(dirname(__DIR__).'/.env');
+// Test if required parameters are set
+if (0 !== count(array_diff($requiredParameters, array_keys($parameters)))) {
+    throw new RuntimeException(sprintf(
+        'Required parameters are not configured, required params are: %s, configure them in %s',
+        implode(', ', $requiredParameters),
+        $parametersPath
+    ));
 }
 
-$_SERVER += $_ENV;
-$_SERVER['APP_ENV'] = $_ENV['APP_ENV'] = ($_SERVER['APP_ENV'] ?? $_ENV['APP_ENV'] ?? null) ?: 'dev';
-$_SERVER['APP_DEBUG'] = $_SERVER['APP_DEBUG'] ?? $_ENV['APP_DEBUG'] ?? 'prod' !== $_SERVER['APP_ENV'];
-$_SERVER['APP_DEBUG'] = $_ENV['APP_DEBUG'] = (int) $_SERVER['APP_DEBUG'] || filter_var($_SERVER['APP_DEBUG'], FILTER_VALIDATE_BOOLEAN) ? '1' : '0';
+$_SERVER['APP_ENV'] = $parameters['app_env'];
+$_SERVER['APP_DEBUG'] = $parameters['app_debug'];
+$_SERVER['APP_SECRET'] = $parameters['app_secret'];
+
+// Allow the application environment (dev/test/prod) to change via the APP_ENV environment variable.
+if (array_key_exists('APP_ENV', $_ENV)) {
+    $_SERVER['APP_ENV'] = $_ENV['APP_ENV'];
+}
+
+filter_var($_SERVER['APP_DEBUG'], FILTER_VALIDATE_BOOLEAN) ? '1' : '0';
